pyproject.toml
[project]
name = "nano-b-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.32.0",
    "ipython>=9.5.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
